+++ FBSDKCoreKit/FBSDKCoreKit/FBSDKApplicationDelegate.m
@@ -28,6 +28,7 @@
 #import "FBSDKAppEventsStateManager+AppEventsStatePersisting.h"
 #import "FBSDKAppEventsUtility+AdvertiserIDProviding.h"
 #import "FBSDKAuthenticationStatusUtility.h"
+#import "FBSDKAuthenticationToken+Internal.h"
 #import "FBSDKBridgeAPI+ApplicationObserving.h"
 #import "FBSDKButton+Subclass.h"
 #import "FBSDKConstants.h"