+++ FBSDKCoreKit/FBSDKCoreKit/FBSDKApplicationObserving.h
@@ -20,6 +20,10 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
+/*
+ Describes any types that optionally responds to various lifecycle events
+ received by the system and propagated by `ApplicationDelegate`.
+ */
 @protocol FBSDKApplicationObserving <NSObject>
 
 @optional